requirements.txt
torch
pandas
numpy
onnxruntime
Pillow
opencv-python
scikit-learn
matplotlib
seaborn
tqdm
torchvision 